+++ deps/lcb/src/n1ql/ixmgmt.cc
@@ -736,7 +736,8 @@ WatchIndexCtx::do_poll()
     lcb_CMDN1XMGMT cmd;
     memset(&cmd, 0, sizeof cmd);
     cmd.callback = cb_watch_gotlist;
-    lcb_log(LOGARGS(this, DEBUG), LOGFMT "Will check for index readiness of %lu indexes. %lu completed", LOGID(this), m_defspend.size(), m_defsok.size());
+    lcb_log(LOGARGS(this, DEBUG), LOGFMT "Will check for index readiness of %lu indexes. %lu completed",
+            LOGID(this), (unsigned long int)m_defspend.size(), (unsigned long int)m_defsok.size());
     return lcb_n1x_list(m_instance, this, &cmd);
 }
